requirements.txt
werkzeug~=3.0.5
pytest~=7.4.4
uvicorn~=0.34.0
python-dotenv~=0.21.0
fastapi~=0.112.2
pydantic~=2.10.3
sqlalchemy~=2.0.37
python-dotenv~=0.21.0
sqlalchemy~=2.0.34
bcrypt~=4.0.1
PyMySQL~=1.1.0
SQLAlchemy~=2.0.34

